make write_to_path more private and at unique call
